crate::shim:consensus
